added create bounties and tooltip
