+++ src/components/checklist/checklist.tsx
@@ -1,38 +1,10 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 "use client";
 
-import { weaponSubmission } from "@/app/schemas/schema";
-import {
-  itemRarity,
-  itemTypes,
-  rarityColors,
-  typeIcons,
-} from "@/lib/item-changes";
 import { api } from "@/lib/trpc/client";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { CheckCircle, Clock, Search, XCircle } from "lucide-react";
+import { Search } from "lucide-react";
 import { useCallback, useEffect, useMemo, useState } from "react";
-import { useForm } from "react-hook-form";
 import { toast } from "sonner";
-import * as z from "zod";
-import { Badge } from "../ui/badge";
-import { Button } from "../ui/button";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "../ui/card";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "../ui/dialog";
-import { Form, FormControl, FormField, FormItem, FormLabel } from "../ui/form";
+import { Card, CardContent } from "../ui/card";
 import { Input } from "../ui/input";
 import {
   Select,
@@ -41,211 +13,7 @@ import {
   SelectTrigger,
   SelectValue,
 } from "../ui/select";
-
-type WeaponSubmissionValues = z.infer<typeof weaponSubmission>;
-
-
-const WeaponCard = ({
-  weapon,
-  onSubmit,
-  isSubmitting,
-}: {
-  weapon: any;
-  onSubmit: (weaponId: string, clipUrl: string) => void;
-  isSubmitting: boolean;
-}) => {
-  const [isDialogOpen, setIsDialogOpen] = useState(false);
-
-  const form = useForm<WeaponSubmissionValues>({
-    resolver: zodResolver(weaponSubmission),
-    defaultValues: {
-      twitchClipLink: "",
-      itemId: weapon.id,
-    },
-  });
-
-  const handleSubmit = useCallback(
-    (data: WeaponSubmissionValues) => {
-      onSubmit(data.itemId, data.twitchClipLink);
-      setIsDialogOpen(false);
-    },
-    [onSubmit]
-  );
-
-  const handleDialogOpen = useCallback(() => {
-    form.setValue("itemId", weapon.id);
-    setIsDialogOpen(true);
-  }, [form, weapon.id]);
-
-  return (
-    <Card className="bg-black/20 border-white/10 text-white relative">
-      <CardHeader>
-        <div className="flex items-center justify-between">
-          <div className="flex-1">
-            <CardTitle className="flex items-center gap-2 text-lg">
-              <span className="text-2xl">
-                {typeIcons[itemTypes(weapon.type)]}
-                {weapon.name}
-              </span>
-            </CardTitle>
-            <CardDescription className="text-gray-300 mt-1">
-              {weapon.source}
-            </CardDescription>
-          </div>
-          <div className="flex flex-col items-end gap-2">
-            <Badge
-              className={`${rarityColors[weapon.rarity]} text-white capitalize`}
-            >
-              {itemRarity(weapon.rarity)}
-            </Badge>
-            {weapon.userItemStatus && (
-              <Badge
-                variant={
-                  weapon.userItemStatus === "Approved"
-                    ? "default"
-                    : weapon.userItemStatus === "Pending"
-                    ? "secondary"
-                    : "destructive"
-                }
-                className="text-xs"
-              >
-                {weapon.userItemStatus === "Approved" && (
-                  <CheckCircle className="h-3 w-3 mr-1" />
-                )}
-                {weapon.userItemStatus === "Pending" && (
-                  <Clock className="h-3 w-3 mr-1" />
-                )}
-                {weapon.userItemStatus === "Rejected" && (
-                  <XCircle className="h-3 w-3 mr-1" />
-                )}
-                {weapon.userItemStatus}
-              </Badge>
-            )}
-          </div>
-        </div>
-      </CardHeader>
-      <CardContent className="space-y-3">
-        <div className="grid gap-2 text-sm">
-          <div>
-            <span>Points:</span>
-            <span className="ml-2 font-semibold text-yellow-400">
-              {weapon.points}
-            </span>
-          </div>
-          <div>
-            <span>Mayhem:</span>
-            <span className="ml-2 font-semibold ">{weapon.mayhem}</span>
-          </div>
-          <div>
-            <span>Type:</span>
-            <span className="ml-2 font-semibold text-yellow-400">
-              {itemTypes(weapon.type)}
-            </span>
-          </div>
-          <div>
-            <span>Group:</span>
-            <span className="ml-2 font-semibold text-yellow-400">
-              {weapon.listGroup}
-            </span>
-          </div>
-          <div>
-            <span>Mission:</span>
-            <span className="ml-2 font-semibold text-yellow-400">
-              {weapon.missionType}
-            </span>
-          </div>
-          <div>
-            <span>Maps:</span>
-            <span className="ml-2 font-semibold text-yellow-400">
-              {weapon.maps}
-            </span>
-          </div>
-        </div>
-        {weapon.notes && (
-          <div>
-            <span className="text-gray-400 text-sm">Notes</span>
-            <p className="text-sm text-gray-300 mt-1">{weapon.notes}</p>
-          </div>
-        )}
-
-        <Dialog open={isDialogOpen} onOpenChange={setIsDialogOpen}>
-          <DialogTrigger asChild>
-            <Button
-              className="w-full mt-4 text-black hover:opacity-90"
-              style={{
-                backgroundColor:
-                  weapon.userItemStatus === "Approved"
-                    ? "rgb(107 114 128)"
-                    : "oklch(0.9181 0.2323 126.72)",
-              }}
-              disabled={
-                weapon.userItemStatus === "Pending" ||
-                weapon.userItemStatus === "Approved" ||
-                isSubmitting
-              }
-              onClick={handleDialogOpen}
-            >
-              {weapon.userItemStatus === "Approved"
-                ? "Completed"
-                : weapon.userItemStatus === "Pending"
-                ? "Pending Review"
-                : weapon.userItemStatus === "Rejected"
-                ? "Resubmit"
-                : isSubmitting
-                ? "Submitting..."
-                : "Submit"}
-            </Button>
-          </DialogTrigger>
-          <DialogContent className="bg-black border-primary-green">
-            <DialogHeader>
-              <DialogTitle className="text-white">
-                Submit {weapon.name}
-              </DialogTitle>
-              <DialogDescription className="text-white">
-                Provide evidence that you&apos;ve obtained this weapon. Include
-                the Twitch clip link.
-              </DialogDescription>
-            </DialogHeader>
-            <div className="space-y-4">
-              <Form {...form}>
-                <form
-                  className="space-y-6"
-                  onSubmit={form.handleSubmit(handleSubmit)}
-                >
-                  <FormField
-                    control={form.control}
-                    name="twitchClipLink"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel className="text-white">
-                          Twitch Clip URL
-                        </FormLabel>
-                        <FormControl>
-                          <Input
-                            {...field}
-                            placeholder="https://www.twitch.tv/[username]/clip/"
-                            className="bg-black border-primary-green text-white focus-visible:ring-primary-green "
-                          />
-                        </FormControl>
-                      </FormItem>
-                    )}
-                  />
-                  <Button
-                    type="submit"
-                    className="bg-primary-green hover:bg-primary-green/50 cursor-pointer text-black"
-                    disabled={isSubmitting}
-                  >
-                    {isSubmitting ? "Submitting..." : "Submit"}
-                  </Button>
-                </form>
-              </Form>
-            </div>
-          </DialogContent>
-        </Dialog>
-      </CardContent>
-    </Card>
-  );
-};
+import { WeaponCard } from "./weapon-card";
 
 export default function Checklist() {
   const [searchQuery, setSearchQuery] = useState("");
@@ -261,6 +29,17 @@ export default function Checklist() {
     },
     onError: (error) => {
       toast.error("Failed to submit evidence. Please try again.");
+      console.error("Submission error:", error);
+    },
+  });
+
+  const bountyMutation = api.bounty.create.useMutation({
+    onSuccess: async () => {
+      toast("Bounty Created");
+      utils.bounty.getAll.invalidate();
+    },
+    onError: (error) => {
+      toast.error("Failed to submit bounty. Please try again.");
       console.error("Submission error:", error);
     },
   });
@@ -304,13 +83,24 @@ export default function Checklist() {
 
   const handleSubmit = useCallback(
     (itemId: string, clipUrl: string) => {
-      console.log(itemId);
       submitMutation.mutate({
         itemId,
         twitchClipUrl: clipUrl,
       });
     },
     [submitMutation]
+  );
+
+  const handleBountySubmit = useCallback(
+    (itemId: string, title: string, price: number, description: string) => {
+      bountyMutation.mutate({
+        itemId,
+        price,
+        title,
+        description,
+      });
+    },
+    [bountyMutation]
   );
 
   const mergedItems = useMemo(() => {
@@ -421,6 +211,7 @@ export default function Checklist() {
             key={weapon.id}
             weapon={weapon}
             onSubmit={handleSubmit}
+            bountySubmit={handleBountySubmit}
             isSubmitting={submitMutation.isPending}
           />
         ))}
